+++ src/main/java/entities/Cleaner.java
@@ -10,7 +10,7 @@ public class Cleaner extends UniversityStaff {
                    String firstName,
                    String lastName,
                    int experience,
-                   int salary) {
+                   double salary) {
         super(faculty, firstName, lastName, experience, salary);
     }
 }
